refactor : 오버플로우 방지

+++ src/main/java/blackjack/domain/participants/betting/Betting.java
@@ -30,7 +30,7 @@ public class Betting {
         return bettingProfit.getProfitResult();
     }
 
-    public int getDealerProfit() {
+    public long getDealerProfit() {
         return bettingProfit.getDealerProfit();
     }
 }
